Extract the LSP functionality into its own module
